Adding authorization with gates (too late this time, sorry)

+++ app/Http/Controllers/Api/AtendeeController.php
@@ -5,13 +5,17 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\AttendeeResource;
 use App\Http\Traits\CanLoadRelationships;
-use App\Models\{Event, Attendee};
+use App\Models\{Event, Attendee, User};
 use Illuminate\Http\Request;
 
 class AtendeeController extends Controller
 {
     use CanLoadRelationships;
     private array $relations = ['user'];
+
+    public function __Construct(){
+        $this->middleware('auth:sanctum')->except(['index','show','update']);
+    }
     public function index(Event $event)
     {
         $attendees = $this->LoadRelationships(
@@ -27,7 +31,7 @@ class AtendeeController extends Controller
     {
         $attendee = $this->LoadRelationships(
             $event->attendees()->create([
-                'user_id' => 1
+                'user_id' => $request->user()->id
             ])
         );
         return new AttendeeResource($attendee);
@@ -38,8 +42,10 @@ class AtendeeController extends Controller
         return new AttendeeResource($this->LoadRelationships($attendee));
     }
 
-    public function destroy(string $event,Attendee $attendee)
+    public function destroy(Event $event,Attendee $attendee)
     {
+        $this->authorize('delete-attendee',[$event,$attendee]);
+
         $attendee->delete();
 
         return response(status: 204);
